requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --output-file=requirements.txt
agentops==0.4.21
    # via personal-notion-agent
agno==1.7.1
    # via personal-notion-agent
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via agentops
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   google-genai
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via aiohttp
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
    #   uvicorn
distlib==0.4.0
    # via virtualenv
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via agno
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.116.1
    # via personal-notion-agent
fastapi-cli==0.0.11
    # via fastapi
fastapi-cloud-cli==0.1.5
    # via fastapi-cli
filelock==3.19.1
    # via virtualenv
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via agno
google-auth==2.40.3
    # via google-genai
google-genai==1.36.0
    # via personal-notion-agent
googleapis-common-protos==1.72.0
    # via opentelemetry-exporter-otlp-proto-http
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   agentops
    #   agno
    #   fastapi
    #   fastapi-cloud-cli
    #   google-genai
    #   notion-client
    #   openai
    #   python-telegram-bot
identify==2.6.14
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via opentelemetry-api
jinja2==3.1.6
    # via fastapi
jiter==0.10.0
    # via openai
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
nodeenv==1.9.1
    # via pre-commit
notion-client==2.5.0
    # via personal-notion-agent
openai==1.107.2
    # via personal-notion-agent
opentelemetry-api==1.38.0
    # via
    #   agentops
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.38.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.38.0
    # via agentops
opentelemetry-instrumentation==0.59b0
    # via agentops
opentelemetry-proto==1.38.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.38.0
    # via
    #   agentops
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.59b0
    # via
    #   agentops
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
ordered-set==4.1.0
    # via agentops
packaging==24.2
    # via
    #   agentops
    #   opentelemetry-instrumentation
platformdirs==4.4.0
    # via virtualenv
pre-commit==4.3.0
    # via personal-notion-agent
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.0
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
psutil==7.0.0
    # via agentops
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   agno
    #   fastapi
    #   fastapi-cloud-cli
    #   google-genai
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via agno
pygments==2.19.2
    # via rich
python-dotenv==1.1.1
    # via
    #   agno
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via
    #   agno
    #   fastapi
python-telegram-bot==22.3
    # via personal-notion-agent
pyyaml==6.0.2
    # via
    #   agentops
    #   agno
    #   pre-commit
    #   uvicorn
requests==2.32.5
    # via
    #   agentops
    #   google-genai
    #   opentelemetry-exporter-otlp-proto-http
    #   personal-notion-agent
rich==14.1.0
    # via
    #   agno
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
rsa==4.7.2
    # via google-auth
ruff==0.13.0
sentry-sdk==2.37.1
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.47.3
    # via fastapi
tenacity==9.1.2
    # via google-genai
termcolor==2.4.0
    # via agentops
tomli==2.2.1
    # via agno
tqdm==4.67.1
    # via openai
typer==0.17.4
    # via
    #   agno
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   agno
    #   aiosignal
    #   anyio
    #   fastapi
    #   google-genai
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.35.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
    #   personal-notion-agent
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
virtualenv==20.34.0
    # via pre-commit
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
wrapt==1.17.3
    # via
    #   agentops
    #   opentelemetry-instrumentation
yarl==1.22.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
